The final runnable jar
Download this file and double click it to run. As stated in the notes the application will create a folder under user/documents/TimeTrackerPro and inside that folder is where the application will create and store the database. Its important this file remains unaltered as that could mess up the applications function and result in data loss.